[add] fn scope warnings to main socpe
