+++ csml_interpreter/src/data/primitive/array.rs
@@ -252,7 +252,7 @@ impl PrimitiveArray {
             return Ok(PrimitiveArray::get_literal(&vector, interval));
         }
 
-        Ok(PrimitiveArray::get_literal( &[], interval))
+        Ok(PrimitiveArray::get_literal(&[], interval))
     }
 
     fn is_empty(
