+++ csml_interpreter/src/interpreter.rs
@@ -1,6 +1,7 @@
 pub mod ast_interpreter;
 pub mod builtins;
 pub mod components;
+pub mod function_scope;
 pub mod json_to_rust;
 pub mod variable_handler;
 
@@ -8,7 +9,7 @@ pub use json_to_rust::{json_to_literal, memory_to_literal};
 
 use crate::data::error_info::ErrorInfo;
 use crate::data::position::Position;
-use crate::data::{ast::*, primitive::PrimitiveNull, Data, Hold, Literal, MessageData, MSG};
+use crate::data::{ast::*, Data, Hold, Literal, MessageData, MSG};
 use crate::error_format::*;
 use crate::interpreter::{
     ast_interpreter::{for_loop, match_actions, solve_if_statement},
@@ -120,60 +121,4 @@ pub fn interpret_scope(
     }
 
     Ok(message_data)
-}
-
-pub fn interpret_function_scope(
-    actions: &Block,
-    data: &mut Data,
-    interval: Interval,
-) -> Result<Literal, ErrorInfo> {
-    let mut message_data = MessageData::default();
-
-    for (action, instruction_info) in actions.commands.iter() {
-        match action {
-            Expr::ObjectExpr(ObjectType::Return(var)) => {
-                let lit = expr_to_literal(var, false, None, data, &mut message_data, &None)?;
-
-                return Ok(lit);
-            }
-            Expr::ObjectExpr(fun) => {
-                message_data = match_actions(fun, message_data, data, None, &None)?
-            }
-            Expr::IfExpr(ref if_statement) => {
-                message_data = solve_if_statement(
-                    if_statement,
-                    message_data,
-                    data,
-                    &None,
-                    instruction_info,
-                    &None,
-                )?;
-            }
-            Expr::ForEachExpr(ident, i, expr, block, range) => {
-                message_data = for_loop(
-                    ident,
-                    i,
-                    expr,
-                    block,
-                    range,
-                    message_data,
-                    data,
-                    &None,
-                    &None,
-                )?
-            }
-            e => {
-                return Err(gen_error_info(
-                    Position::new(interval_from_expr(e)),
-                    ERROR_START_INSTRUCTIONS.to_owned(),
-                ));
-            }
-        };
-
-        if let Some(ExitCondition::Return(lit)) = message_data.exit_condition {
-            return Ok(lit);
-        }
-    }
-
-    Ok(PrimitiveNull::get_literal(interval))
 }
